+++ js/form-validation.js
@@ -1,5 +1,10 @@
+const hashtagInput = document.querySelector('.text__hashtags');
+const commentInput = document.querySelector('.text__description');
+let message;
+
 const testStartWith = (hashtag) => {
   if (!hashtag.startsWith('#')) {
+    message = 'хэш-тег должен начинаться с символа # (решётка)';
     return 'хэш-тег должен начинаться с символа # (решётка)';
   }
   return undefined;
@@ -7,6 +12,7 @@ const testStartWith = (hashtag) => {
 
 const testShortValueLength = (hashtag) => {
   if (hashtag.length === 1) {
+    message = 'хеш-тег не может состоять только из одной решётки';
     return 'хеш-тег не может состоять только из одной решётки';
   }
   return undefined;
@@ -16,6 +22,7 @@ const testValidity = (hashtag) => {
   const regex = /^[A-Za-z0-9]+$/;
   const isValid = regex.test(hashtag.split('#')[1]);
   if (!isValid) {
+    message = 'хэш-тэг не может содержать пробелы, спецсимволы (#, @, $ и т. п.), символы пунктуации (тире, дефис, запятая и т. п.), эмодзи и т. д.';
     return 'хэш-тэг не может содержать пробелы, спецсимволы (#, @, $ и т. п.), символы пунктуации (тире, дефис, запятая и т. п.), эмодзи и т. д.';
   }
   return undefined;
@@ -23,6 +30,7 @@ const testValidity = (hashtag) => {
 
 const testLongValueLength = (hashtag) => {
   if (hashtag.length > 20) {
+    message = 'максимальная длина одного хэш-тега 20 символов, включая решётку';
     return 'максимальная длина одного хэш-тега 20 символов, включая решётку';
   }
   return undefined;
@@ -30,6 +38,7 @@ const testLongValueLength = (hashtag) => {
 
 const testUniqueName = (hashtagArray, index) => {
   if (hashtagArray[index - 1] === hashtagArray[index]) {
+    message = 'один и тот же хэш-тег не может быть использован дважды';
     return 'один и тот же хэш-тег не может быть использован дважды';
   }
   return undefined;
@@ -37,6 +46,7 @@ const testUniqueName = (hashtagArray, index) => {
 
 const testHashtagQuantity = (hashtagArray) => {
   if (hashtagArray.length > 5) {
+    message = 'нельзя указать больше пяти хэш-тегов';
     return 'нельзя указать больше пяти хэш-тегов';
   }
   return undefined;
@@ -55,6 +65,48 @@ const highlightErrorBackground = (element) => {
 
 const whitenBackground = (element) => {
   element.style.backgroundColor = 'white';
-}
+};
 
-export {testStartWith, testShortValueLength, testValidity, testLongValueLength, testUniqueName, testHashtagQuantity, testCommentLength, highlightErrorBackground, whitenBackground};
+const testHashtagInput = () => {
+  const hashtagArray = hashtagInput.value.toLowerCase().split(' ');
+
+  let error = hashtagArray.some((hashtag, index) => {
+    return (
+      testStartWith(hashtag)
+      || testShortValueLength(hashtag)
+      || testValidity(hashtag)
+      || testLongValueLength(hashtag)
+      || testUniqueName(hashtagArray, index)
+      || testHashtagQuantity(hashtagArray)
+    );
+  });
+
+  if (error) {
+    highlightErrorBackground(hashtagInput);
+    hashtagInput.setCustomValidity(message);
+  } else {
+    whitenBackground(hashtagInput);
+    hashtagInput.setCustomValidity('');
+  }
+  hashtagInput.reportValidity();
+
+  if (hashtagInput.value === '') {
+    whitenBackground(hashtagInput);
+    hashtagInput.setCustomValidity('');
+  }
+  hashtagInput.reportValidity();
+};
+
+const testCommentInput = () => {
+  let error = testCommentLength(commentInput);
+  if (error) {
+    highlightErrorBackground(commentInput);
+    commentInput.setCustomValidity(error);
+  } else {
+    whitenBackground(commentInput);
+    commentInput.setCustomValidity('');
+  }
+  commentInput.reportValidity();
+};
+
+export {hashtagInput, commentInput, testHashtagInput, testCommentInput, whitenBackground};
