+++ js/util.js
@@ -1,4 +1,7 @@
 const body = document.querySelector('body');
+const main = document.querySelector('main');
+const alertSuccessTemplate = document.querySelector('#success').content.querySelector('.success');
+const alertErrorTemplate = document.querySelector('#error').content.querySelector('.error');
 
 const isEscEvent = (evt) => {
   return evt.key === ('Escape' || 'Esc');
@@ -8,77 +11,60 @@ const isEnterEvent = (evt) => {
   return evt.key === 'Enter';
 };
 
-const commentList = document.querySelector('.social__comments');
-/*const commentField = document.querySelector('.social__footer-text');
-const commentBtn = document.querySelector('.social__footer-btn');
+const findSuccessButton = () => {
+  const successButton = main.querySelector('.success__button');
+  return successButton;
+};
 
-commentBtn.onclick = (evt) => {
-  evt.preventDefault();
+const findErrorButton = () => {
+  const errorButton = main.querySelector('.error__button');
+  return errorButton;
+};
 
-  const newComment = document.createElement('li');
-  newComment.classList.add('social__comment');
+const hideSuccessMessage = () => {
+  body.classList.remove('modal-open');
 
-  const newCommentImg = document.createElement('img');
-  newCommentImg.classList.add('social__picture');
-  newCommentImg.src = document.querySelector('.big-picture').querySelector('.social__footer').querySelector('.social__picture').src;
-  newCommentImg.alt = 'имя комментатора';
-  newCommentImg.width = 35;
-  newCommentImg.height = 35;
-  newComment.appendChild(newCommentImg);
+  findSuccessButton().removeEventListener('click', hideSuccessMessage);
+  document.removeEventListener('keydown', onSuccessMessageEscKeydown);
+  document.removeEventListener('mouseup', onSuccessMessageMouseUp);
 
-  const newCommentText = document.createElement('p');
-  newCommentText.classList.add('social__text');
-  newComment.appendChild(newCommentText);
-  newCommentText.textContent = commentField.value;
+  main.querySelector('.success').remove();
+};
 
-  if (commentField.value === '') {
-    alert('Комментарии не могут быть пустыми – напишите что-нибудь');
-  } else {
-    commentField.value = '';
-    commentList.append(newComment);
-  }
-};*/
+const hideErrorMessage = () => {
+  body.classList.remove('modal-open');
 
-const main = document.querySelector('main');
-const alertSuccessTemplate = document.querySelector('#success').content.querySelector('.success');
-const alertErrorTemplate = document.querySelector('#error').content.querySelector('.error');
+  findErrorButton().removeEventListener('click', hideErrorMessage);
+  document.removeEventListener('keydown', onErrorMessageEscKeydown);
+  document.removeEventListener('mouseup', onErrorMessageMouseUp);
+
+  main.querySelector('.error').remove();
+};
 
 const onSuccessMessageEscKeydown = (evt) => {
   if (isEscEvent(evt)) {
     evt.preventDefault();
-    main.querySelector('.success').remove();
-    document.removeEventListener('keydown', onSuccessMessageEscKeydown);
+    hideSuccessMessage();
   }
 };
 
 const onErrorMessageEscKeydown = (evt) => {
   if (isEscEvent(evt)) {
     evt.preventDefault();
-    main.querySelector('.error').remove();
-    document.removeEventListener('keydown', onErrorMessageEscKeydown);
+    hideErrorMessage();
   }
 };
 
 const onSuccessMessageMouseUp = (evt) => {
-  if (!main.querySelector('.success__inner').contains(evt.target)) {
-    main.querySelector('.success').remove();
-    document.removeEventListener('mouseup', onSuccessMessageMouseUp);
+  if (evt.target !== main.querySelector('.success__inner')) {
+    hideSuccessMessage();
   }
 };
 
 const onErrorMessageMouseUp = (evt) => {
-  if (!main.querySelector('.error__inner').contains(evt.target)) {
-    main.querySelector('.error').remove();
-    document.removeEventListener('mouseup', onErrorMessageMouseUp);
+  if (evt.target !== main.querySelector('.error__inner')) {
+    hideErrorMessage();
   }
-};
-
-const hideSuccessMessage = () => {
-  main.querySelector('.success').remove();
-};
-
-const hideErrorMessage = () => {
-  main.querySelector('.error').remove();
 };
 
 const showSuccessMessage = () => {
@@ -87,13 +73,12 @@ const showSuccessMessage = () => {
   alertBox.appendChild(successSection);
   main.appendChild(alertBox);
 
-  const successButton = main.querySelector('.success__button');
+  body.classList.add('modal-open');
 
-  successButton.addEventListener('click', hideSuccessMessage)
-
+  findSuccessButton().addEventListener('click', hideSuccessMessage);
   document.addEventListener('keydown', onSuccessMessageEscKeydown);
   document.addEventListener('mouseup', onSuccessMessageMouseUp);
-}
+};
 
 const showErrorMessage = () => {
   const alertBox = document.createDocumentFragment();
@@ -101,12 +86,11 @@ const showErrorMessage = () => {
   alertBox.appendChild(errorSection);
   main.appendChild(alertBox);
 
-  const errorButton = main.querySelector('.error__button');
+  body.classList.add('modal-open');
 
-  errorButton.addEventListener('click', hideErrorMessage)
-
+  findErrorButton().addEventListener('click', hideErrorMessage);
   document.addEventListener('keydown', onErrorMessageEscKeydown);
   document.addEventListener('mouseup', onErrorMessageMouseUp);
-}
+};
 
-export {body, showSuccessMessage, showErrorMessage, isEnterEvent, isEscEvent, commentList};
+export {body, showSuccessMessage, showErrorMessage, isEnterEvent, isEscEvent};
